requirement.txt
google-genai
pyaudio
SpeechRecognition